Fix: 기존의 diaryRepository.countByMemberId 로 친구 명단 전체 조회 후 List의 개수를 세는 방식으로 수정 → 추후 리펙토링 예정

+++ src/main/java/org/lxdproject/lxd/diary/dto/MemberDiarySummaryResponseDTO.java
@@ -11,6 +11,6 @@ public class MemberDiarySummaryResponseDTO {
     private String username;
     private String nickname;
     private Long diaryCount;
-    private Long friendCount;
+    private Integer friendCount;
     private RelationType relation;
 }
